refactor(dashboard-tsr): split assistant-ui thread.tsx into -thread/* (#1392)
thread.tsx 1107 -> 399. Extract co-located modules:
- -thread/composer.tsx: welcome + in-thread mention composers
- -thread/chain-of-thought.tsx: GroupedParts reasoning/tool pipeline
- -thread/message-stats.tsx: per-message stats footer + details dialog
- -thread/format.ts: shared duration/date helpers (now wired in)

Behavior-preserving; tsc clean.
